requirements.txt
django>=5.0,<5.1
djangorestframework>=3.15.0,<3.16
drf-spectacular>=0.27.0,<0.28
django-environ>=0.11.0,<0.12
psycopg2-binary>=2.9.9,<2.10
gunicorn>=21.2.0,<21.3

